--- custom_events/src/Components/Buttons/AddCalBtn.js
@@ -1,30 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import { withStyles } from '@material-ui/core/styles';
-import Button from '@material-ui/core/Button';
-
-const styles = theme => ({
-  button: {
-    margin: theme.spacing.unit,
-  },
-  input: {
-    display: 'none',
-  },
-});
-
-function ContainedButtons(props) {
-  const { classes } = props;
-  return (
-    <div>
-      <Button variant="contained" color="primary" className={classes.button}>
-        Add to Calendar
-      </Button>
-    </div>
-  );
-}
-
-ContainedButtons.propTypes = {
-  classes: PropTypes.object.isRequired,
-};
-
-export default withStyles(styles)(ContainedButtons);
